+++ app/src/main/java/com/example/restcountriesapp/feature/countries/navigation/CountriesNavigationKey.kt
@@ -1,0 +1,8 @@
+package com.example.restcountriesapp.feature.countries.navigation
+
+
+data object CountriesListKey
+
+data class CountryDetailsKey(
+    val countryCode: String
+)